requirements.txt
Flask==2.2.5
Pillow==9.5.0
arabic-reshaper==3.0.0
python-bidi==0.4.2
gunicorn==20.1.0
